examen/src/test/java/com/example/examen/controller/PrecioControllerTest.java
package com.example.examen.controller;

import java.util.List;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import com.example.examen.model.Precio;
import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNotNull;

@ExtendWith(MockitoExtension.class)
public class PrecioControllerTest {
	@InjectMocks
	private PrecioController precioController;
	@Test
	public void test() {
		ResponseEntity<List<Precio>> responseEntity = precioController.getPrecios();
		assertEquals(HttpStatus.OK,responseEntity.getStatusCode());
		assertNotNull(responseEntity.getBody());
	}

}
